% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lm_table.R
\name{build_lm_table}
\alias{build_lm_table}
\title{Interactively Make Landmark Synonymy Table}
\usage{
build_lm_table(target_spec, original_spec, description, sub_num, begin = 1)
}
\arguments{
\item{target_spec}{a p x 3 matrix of 3d landmark data for the template or specimen that has the correct landmarking scheme}

\item{original_spec}{a matrix of 3d landmarks representing the landmarking scheme to be changed}

\item{description}{a character vector giving names or descriptions for each landmark in the target template}

\item{sub_num}{a numerical vector giving numbers for each landmark in the target template, ideally per-curve or per-patch}
}
\value{
a table
}
\description{
Interactively Make Landmark Synonymy Table
}
\examples{
my_lm_table<-build_lm_table(target_spec = bird_lm_curves[,,1],
original_spec = croclms[,,1],
description = birdmoduledefs$description,
sub_num = birdmoduledefs$landmark_designation,
begin=1)
}
