% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.curves.R
\name{check.curves}
\alias{check.curves}
\title{Utility function to test whether there are any curves missing from specimens}
\usage{
check.curves(lm.tibble)
}
\arguments{
\item{lm.tibble}{a tibble of landmark data}
}
\value{
just used inside of \code{import_chkpt_data} really
}
\description{
Utility function to test whether there are any curves missing from specimens
}
