% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.modules.R
\name{check.modules}
\alias{check.modules}
\title{A version of the checkLM function from Morpho that plots specimens by module identity}
\usage{
check.modules(
  dat.array,
  path = NULL,
  prefix = "",
  suffix = ".ply",
  col = "white",
  pt.size = NULL,
  alpha = 1,
  begin = 1,
  render = c("w", "s"),
  point = c("s", "p"),
  add = FALSE,
  meshlist = NULL,
  Rdata = FALSE,
  module.colors = NULL,
  text.lm = FALSE
)
}
\arguments{
\item{lm.tibble}{a tibble of landmark data}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
A version of the checkLM function from Morpho that plots specimens by module identity
}
\examples{
color.palette <- c( "#6a3d9a","dimgrey","#fb9a99",  "gold", "#009E73",  "#D55E00", "#CC79A7", "cyan2",  "#e31a1c", "#0072B2", "#b2df8a", "#E69F00",  "whitesmoke" ,  "deeppink",   "#a6cee3",   "#F0E442","blue","red","brown", "black")
modulecolors1<-as.factor(module_defs$posthoc.jug.sep)
levels(modulecolors1)<-color.palette
check.modules(patched,path="./ply/",suffix=".ply",pt.size=.5,render="s",module.colors = modulecolors1,begin=1)
}
