% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_curve_info.R
\name{create_curve_info}
\alias{create_curve_info}
\title{Create curve info object}
\usage{
create_curve_info(curvedata, n_fixed)
}
\arguments{
\item{curvedata}{a dataframe or tibble with columns "lm1", "lm2" and "ptswanted".}

\item{n_fixed}{a number indicating how many landmarks are fixed anatomical landmarks}
}
\value{
output contains the necessary information to subsample curves, and the identity of each curve for sliding
\item{Curves}{a list of curves with the landmarks that begin and end each curve}
\item{Curve.in}{the same list of curves without the landmarks that begin and end each}
\item{Sliding.LMs}{an integer vector of all the points that are sliding semilandmarks on curves}
\item{Fixed}{an integer vector of all the points that are fixed landmarks that are not allowed to slide}
}
\description{
Create curve info object
}
\examples{
my_curvedata <- read_csv("curvedatafile.csv")
my_curves <- create_curve_info(curvedata = my_curvedata, n_fixed = 32)
}
