% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_chkpt_data.R
\name{import_chkpt_data}
\alias{import_chkpt_data}
\title{Import PTS files from Checkpoint}
\usage{
import_chkpt_data(ptslist, curveinfo, subsampl = TRUE)
}
\arguments{
\item{ptslist}{a character vector of file names indicating the pts files to be imported}

\item{curveinfo}{the output of the function \code{create_curve_info}.}

\item{subsample}{a logical value specifying whether or not to subsample the semilandmark curves to equal lengths}
}
\value{
If \code{subsample = TRUE}, returns a 3D array containing imported XYZ coordinate data, compatible with geomorph analyses. If \code{subsample = FALSE}, returns a tibble of coordinate data.
}
\description{
Import PTS files from Checkpoint
}
\examples{
my_curvedata <- read_csv("curvedatafile.csv") #read data
my_curves <- create_curve_info(curvedata = my_curvedata, n_fixed = 32)
import_chkpt_data(ptslist = filenames, curveinfo = my_curves, subsampl = TRUE)
}
