% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{market_data}
\alias{market_data}
\title{Historical Alberta Energy Market Data}
\format{
A data frame with 2 variables:
\describe{
\item{price}{The actual hourly pool price in CAD/MWh}
\item{load}{The Alberta Internal Load (AIL) in MW}
}
}
\source{
\url{http://ets.aeso.ca/}
}
\usage{
market_data
}
\description{
A dataset containing historical hourly pool prices and internal load (demand)
for Alberta. Used to calculate risk probabilities.
}
\keyword{datasets}
