% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_eia.R
\name{pull_eia}
\alias{pull_eia}
\title{Pulls data from the EIA API as a tibble}
\usage{
pull_eia(api_key, route, data = "value", facets = NULL)
}
\arguments{
\item{api_key}{Your EIA API key}

\item{route}{API endpoint route}

\item{data}{name of value column to return, "value" is default}

\item{facets}{Name of data columns to fetch}
}
\value{
A tibble with API data
}
\description{
Fetches data from the EIA API v2.
Constructs the request and query parameters, and parses JSON into a tibble.
Stops on API errors.
}
\examples{
\dontrun{
Need a valid API key to run this
key <- "api_key"
petroleum_stocks <- pull_eia(
  api_key = key,
  route = "petroleum/stoc/wstk/data/",
  facets = list(series = c("WTESTUS1", "WTTSTUS1")))
}
}
